@@ -6,61 +6,63 @@
     "test": "karma start ClientApp/test/karma.conf.js"
   },
   "devDependencies": {
-    "@angular/animations": "4.3.1",
-    "@angular/common": "4.3.1",
-    "@angular/compiler": "4.3.1",
-    "@angular/compiler-cli": "4.3.1",
-    "@angular/core": "4.3.1",
-    "@angular/forms": "4.3.1",
-    "@angular/http": "4.3.1",
-    "@angular/platform-browser": "4.3.1",
-    "@angular/platform-browser-dynamic": "4.3.1",
-    "@angular/platform-server": "4.3.1",
-    "@angular/router": "4.3.1",
-    "@ngtools/webpack": "1.5.0",
-    "@types/chai": "4.0.1",
-    "@types/jasmine": "2.5.53",
-    "@types/webpack-env": "1.13.0",
+    "@angular/animations": "6.0.1",
+    "@angular/common": "6.0.1",
+    "@angular/compiler": "6.0.1",
+    "@angular/compiler-cli": "6.0.1",
+    "@angular/core": "6.0.1",
+    "@angular/forms": "6.0.1",
+    "@angular/http": "6.0.1",
+    "@angular/platform-browser": "6.0.1",
+    "@angular/platform-browser-dynamic": "6.0.1",
+    "@angular/platform-server": "6.0.1",
+    "@angular/router": "6.0.1",
+    "@ngtools/webpack": "6.0.1",
+    "@types/chai": "4.1.3",
+    "@types/jasmine": "2.8.7",
+    "@types/webpack-env": "1.13.6",
     "angular2-router-loader": "0.3.5",
     "angular2-template-loader": "0.6.2",
     "aspnet-prerendering": "^3.0.1",
     "aspnet-webpack": "^2.0.1",
-    "awesome-typescript-loader": "3.2.1",
-    "bootstrap": "3.3.7",
-    "chai": "4.0.2",
-    "css": "2.2.1",
-    "css-loader": "0.28.4",
+    "awesome-typescript-loader": "5.0.0",
+    "bootstrap": "4.1.1",
+    "chai": "4.1.2",
+    "css": "2.2.3",
+    "css-loader": "0.28.11",
     "es6-shim": "0.35.3",
-    "event-source-polyfill": "0.0.9",
-    "expose-loader": "0.7.3",
-    "extract-text-webpack-plugin": "2.1.2",
-    "file-loader": "0.11.2",
-    "html-loader": "0.4.5",
+    "event-source-polyfill": "0.0.12",
+    "expose-loader": "0.7.5",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.11",
+    "html-loader": "0.5.5",
     "isomorphic-fetch": "2.2.1",
-    "jasmine-core": "2.6.4",
-    "jquery": "3.2.1",
-    "json-loader": "0.5.4",
-    "karma": "1.7.0",
+    "jasmine-core": "3.1.0",
+    "jquery": "3.3.1",
+    "json-loader": "0.5.7",
+    "karma": "2.0.2",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
-    "karma-jasmine": "1.1.0",
-    "karma-webpack": "2.0.3",
-    "preboot": "4.5.2",
+    "karma-jasmine": "1.1.2",
+    "karma-webpack": "3.0.0",
+    "preboot": "6.0.0-beta.4",
     "raw-loader": "0.5.1",
-    "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.2",
-    "style-loader": "0.18.2",
+    "reflect-metadata": "0.1.12",
+    "rxjs": "6.1.0",
+    "rxjs-compat": "6.1.0",
+    "style-loader": "0.21.0",
     "to-string-loader": "1.1.5",
-    "typescript": "2.4.1",
-    "url-loader": "0.5.9",
-    "webpack": "2.5.1",
-    "webpack-hot-middleware": "2.18.2",
-    "webpack-merge": "4.1.0",
-    "zone.js": "0.8.12"
+    "typescript": "2.7.2",
+    "url-loader": "1.0.1",
+    "webpack": "4.8.3",
+    "webpack-hot-middleware": "2.22.2",
+    "webpack-merge": "4.1.2",
+    "zone.js": "0.8.26",
+    "popper.js": "^1.14.3"
   },
   "dependencies": {
-    "adal-angular4": "1.1.11",
+    "adal-angular4": "3.0.1",
     "ngx-bootstrap": "^2.0.3"
   }
 }
