@@ -1,6 +1,6 @@
 ﻿import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { Adal4Service } from 'adal-angular4';
+import { AdalService } from 'adal-angular4';
 import { AuthHelperService } from '../../services/auth-helper.service';
 
 @Component({
@@ -10,7 +10,7 @@ import { AuthHelperService } from '../../services/auth-helper.service';
 })
 export class ExportComponent {
 
-    constructor(private readonly router: Router, private readonly adalService: Adal4Service,
+    constructor(private readonly router: Router, private readonly adalService: AdalService,
         private readonly authHelper: AuthHelperService) {
     }
 
